Move test to another module
