+++ test-app/build.gradle
@@ -15,7 +15,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.fabenovel.statefullayout.sample"
+        applicationId "com.fabenovel.statefullayout.test"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
@@ -53,4 +53,16 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.2.0'
     implementation 'androidx.fragment:fragment-ktx:1.2.4'
     implementation(project(":lib"))
+
+    testImplementation 'junit:junit:4.13'
+    androidTestImplementation 'androidx.test:core:1.2.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+
+    testImplementation 'androidx.test:core:1.2.0'
+    testImplementation 'androidx.test.ext:junit:1.1.1'
+    testImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+    testImplementation "org.robolectric:robolectric:4.3.1"
+    testImplementation(group: 'androidx.test.espresso', name: 'espresso-contrib', version: '3.1.0-alpha3')
+    debugImplementation 'androidx.fragment:fragment-testing:1.2.4'
 }
