+++ test-app/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Jun 04 09:17:22 CEST 2020
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-6.1.1-all.zip
