+++ test-app/src/main/java/com/fabernovel/statefullayout/test/MainActivity.kt
@@ -1,0 +1,18 @@
+package com.fabernovel.statefullayout.test
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.fabernovel.statefullayout.test.ui.main.MainFragment
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.main_activity)
+        if (savedInstanceState == null) {
+            supportFragmentManager.beginTransaction()
+                .replace(R.id.container, MainFragment.newInstance())
+                .commitNow()
+        }
+    }
+}
