+++ test-app/src/main/java/com/fabernovel/statefullayout/test/ui/empty/EmptyFragment.kt
@@ -1,0 +1,6 @@
+package com.fabernovel.statefullayout.test.ui.empty
+
+import androidx.fragment.app.Fragment
+import com.fabernovel.statefullayout.test.R
+
+class EmptyFragment : Fragment(R.layout.empty_fragment)
