+++ test-app/src/main/java/com/fabernovel/statefullayout/test/ui/override/OverrideDefaultFragment.kt
@@ -1,0 +1,6 @@
+package com.fabernovel.statefullayout.test.ui.override
+
+import androidx.fragment.app.Fragment
+import com.fabernovel.statefullayout.test.R
+
+class OverrideDefaultFragment : Fragment(R.layout.override_default_fragment)
