+++ test-app/src/main/res/layout/empty_fragment.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<com.fabernovel.statefullayout.StatefulLayout
+    android:id="@+id/empty_stateful_layout"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"/>
