+++ test-app/src/main/res/layout/override_default_fragment.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="utf-8"?>
+<com.fabernovel.statefullayout.StatefulLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/stateful_layout"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    >
+
+    <com.fabernovel.statefullayout.State
+        android:id="@+id/stateLoading"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        >
+        <TextView
+            android:id="@+id/custom_loading_text"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:text="@string/custom_loading_state"
+            />
+    </com.fabernovel.statefullayout.State>
+
+    <com.fabernovel.statefullayout.State
+        android:id="@+id/stateError"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        >
+
+        <TextView
+            android:id="@+id/custom_error_text"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:text="@string/custom_error_state"
+            />
+    </com.fabernovel.statefullayout.State>
+</com.fabernovel.statefullayout.StatefulLayout>
