+++ test-app/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<resources>
+    <string name="app_name">test-app</string>
+    <string name="custom_error_state">Custom error state</string>
+    <string name="custom_loading_state">Custom loading state</string>
+</resources>
