+++ test-app/src/test/java/com/fabernovel/statefullayout/utils/clicks.kt
